Fixed require paths for relocated unit test files
