+++ ghost/core/test/unit/server/services/members/members-api/services/GeolocationService.test.js
@@ -1,6 +1,6 @@
 const nock = require('nock');
 const should = require('should');
-const GeolocationService = require('../../../../../../core/server/services/members/members-api/services/GeolocationService');
+const GeolocationService = require('../../../../../../../core/server/services/members/members-api/services/GeolocationService');
 
 const RESPONSE = {
     longitude: '-2.2417',
